Digestive_system_diagram_de.svg‎  (SVG file, nominally 633 × 801 pixels, file size: 207 KB) Description 	Digestive system diagram de.svg
The gastrointestinal tract, also called the digestive tract, alimentary canal, or gut, is the system of organs within multicellular animals that takes in food, digests it to extract energy and nutrients, and expels the remaining waste.
Source

Self made
Date

August, 28 2007
Author

eingedeutscht User:BMK / LadyofHats
Permission
(Reusing this image)
	This image has been released into the public domain by its author, LadyofHats. This applies worldwide.

In some countries this may not be legally possible; if so:
LadyofHats grants anyone the right to use this work for any purpose, without any conditions, unless such conditions are required by law.