CI: Bugfix "Only allow merge requests to be merged if the pipeline succeeds"
When this setting is enabled, a merge request is prevented from being merged if there is no pipeline. This may conflict with some use cases where only/except rules are used and they don’t generate any pipelines.

Users that expect to be able to merge a merge request in this scenario should ensure that there is always a pipeline and that it’s succesful.

For example, to that on merge requests there is always a passing job even though only/except rules may not generate any other jobs: